fix : DTO, data layer

+++ app/src/main/java/com/mate/carpool/data/model/DTO/MemberResponseDTO.kt
@@ -1,7 +1,7 @@
 package com.mate.carpool.data.model.DTO
 
 data class MemberResponseDTO(
-    val passengerId:Int,
+    val passengerId:Long,
     val email:String,
     val studentNumber:String,
     val memberName:String,
