+++ app/src/main/java/com/mate/carpool/data/model/DTO/TicketNewMemberRequestDTO.kt
@@ -1,5 +1,5 @@
 package com.mate.carpool.data.model.DTO
 
 data class TicketNewMemberRequestDTO(
-    val ticketId:Int
+    val ticketId:Long
 )
